pyproject.toml
[tool.bumpver]
current_version = "2026.2.7"
version_pattern = "YYYY.MM.INC1"
commit_message = ":bookmark: bump version {old_version} -> {new_version} [skip ci]"
commit = true
push = false  # disabled for GitHub Actions
tag = true

[tool.bumpver.file_patterns]
"core/__init__.py" = [
    '__version__ = "{version}"',
]

[tool.pytest.ini_options]
addopts = "--durations=1"
DJANGO_SETTINGS_MODULE = "tests.settings"
norecursedirs = ".git config node_modules scss sponsors static templates"
python_files = "test_*.py"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501", "E741"] # temporary

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "CONTRIBUTORS.md",
    "dist",
    "migrations",
    "node_modules",
    "static",
    "venv",
]

per-file-ignores = {}

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.12
target-version = "py312"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[project]
name="djangopackages"
version="2025.10.10"
requires-python = ">=3.12"
dependencies = [
    "cloudflare",
    "croniter",
    "django-anymail",
    "django-better-admin-arrayfield",
    "django-click",
    "django-debug-toolbar",
    "django-environ",
    "django-extensions",
    "django-health-check",
    "django-htmx",
    "django-jsonview",
    "django-maintenance-mode",
    "django-q2",
    "django-reversion",
    "django-secure",
    "django-structlog",
    "django-tailwind-cli>=4.4.2",
    "django-waffle",
    "Django>5.0",
    "djangorestframework<3.16",
    "emoji",
    "feedparser",
    # github3.py
    "github3.py<4.1",
    "gunicorn",
    "looseversion",
    "markdown-it-py[linkify]",
    "markdownify",
    "numpy",
    "openai",
    "packaging",
    "psutil",
    "psycopg[binary, pool]",
    "pydantic",
    "PyGithub",
    "pypistats",
    "python-dateutil",
    "python-gitlab",
    "pytz",
    "requests",
    "requests-mock",
    "rich",
    "sentry-sdk",
    "social-auth-app-django==5.4.3",
    "social-auth-core",
    "sqlite-utils",
    "structlog",
    "trove-classifiers",
    "Unidecode",
    "uwsgi",
    "whitenoise",
    # Redis support
    "redis",
    "blessed>=1.25.0",
]

[dependency-groups]
dev = [
    # testing
    "dirty-equals",
    "django-test-plus",
    "Faker",
    "model-bakery",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "time-machine",

    # pre-commit - please note all linters and autoformatters are in our
    # .pre-commit-config.yaml file
    "pre-commit",
]

ci = [
    # utils
  "bumpver",
]

docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings[python]",
]

maintenance = [
  "PyGithub==2.6.1",
  "jinja2==3.1.6",
]
